Almost fixed prior problems
1. We are chopping images in non interger coordinates.
A: Didn't make much of a difference but helps to get rid of gaps if we
stick to integers.

2. The aspect ratio of the image segment and the segment do not match.
A: This was a big issue, the aspect ratio of the segment had become much
shorter and wider than the imgSegment and as such the displayed image
segment was not displaying its whole width.

3. For some reason the end of the wheel is not aligning properly.
A: I was using 365 as the number of degrees in a full circle.

Note: We now have just half a kitty on the wheel, could be cool to use
this to have a kind of sliding kitty that eventually matches up.