Added a front panel, doesn't look that great
The front panel is more 3d stuff but since we cant shade it it looks
pretty bad. We could pseudo shade it maybe by writing script that takes
into account its angle but that is a bit of a hassle. Instead Im now
thinking to just add a mask that looks nice and has some viewports.